Projet_stage/requirements.txt
addict==2.4.0
adjustText==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
boltons==25.0.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
choreographer==1.1.1
click==8.3.0
colorama==0.4.6
colourmap==1.2.0
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.3
coverage==7.10.7
cryptography==46.0.1
cycler==0.12.1
dash==3.2.0
datazets==1.1.3
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
dnspython==2.8.0
docutils==0.21.2
doxypypy==0.8.8.7
email-validator==2.3.0
executing==2.2.1
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.0
frozenlist==1.8.0
fsspec==2025.9.0
gherkin-official==35.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
iniconfig==2.1.0
invoke==2.2.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaleido==1.1.0
kiwisolver==1.4.9
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.0
logistro==1.1.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
narwhals==2.5.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
open3d==0.19.0
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
pandas==2.3.2
paramiko==4.0.0
parse==1.20.2
parso==0.8.5
patsy==1.0.1
pca==2.10.0
pdoc==15.0.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
plumbum==1.9.0
prompt_toolkit==3.0.52
propcache==0.4.0
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydicom==3.0.1
pydoctest==0.2.1
Pygments==2.19.2
PyNaCl==1.6.0
pynndescent==0.5.13
pyparsing==3.2.4
pyquaternion==0.9.9
pytest==8.4.2
pytest-cov==7.0.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
retrying==1.4.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
roman-numerals-py==3.1.0
rpds-py==0.27.1
rpyc==6.0.2
scatterd==1.4.1
schema==0.7.7
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
sympy==1.14.0
termcolor==3.1.0
terminaltables==3.1.10
testplan==25.8.0
threadpoolctl==3.6.0
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
tsimcne==0.4.24
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap==0.1.1
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
validators==0.20.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
yarl==1.22.0
zipp==3.23.0
